+++ internal/usecases/find_executions.go
@@ -16,10 +16,10 @@ type ExecutionDTO struct {
 	JobName  string  `json:"job_name"`
 	Date     string  `json:"date"`
 	Status   string  `json:"status"`
-	MemUsage float32 `json:"mem_usage"`
-	CPUUsage float32 `json:"cpu_usage"`
-	NetIn    float32 `json:"net_in"`
-	NetOut   float32 `json:"net_out"`
+	MemUsage float64 `json:"mem_usage"`
+	CPUUsage float64 `json:"cpu_usage"`
+	NetIn    int     `json:"net_in"`
+	NetOut   int     `json:"net_out"`
 }
 
 // FindExecutions finds last N executions of all jobs or a specific job
